Updating reworked UI flow to the correct URL
