--- src/pages/delete-account-auth0-callback.js
@@ -1,10 +1,0 @@
-import React from 'react';
-import { handleAuthentication } from '../utils/auth';
-
-const Callback = () => {
-  handleAuthentication();
-
-  return <p>Loading...</p>
-};
-
-export default Callback;
